requirements.txt
alembic==1.4.2
autopep8==1.5.2
branca==0.4.1
censusgeocode==0.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
cycler==0.10.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
folium==0.11.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mpld3==0.3
numpy==1.18.4
pandas==1.0.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.2.1
